Fix rounding errors for setting new volume after button-up or scroll-up.
* applet.c: (gnome_volume_applet_scroll),
(gnome_volume_applet_adjust_volume), (cb_volume):
* applet.h:
* dock.c: (cb_timeout):
Fix rounding errors for setting new volume after button-up or
scroll-up. Fixes #316922.

+++ mixer/dock.c
@@ -196,7 +196,7 @@ cb_timeout (gpointer data)
 {
   GnomeVolumeAppletDock *dock = data;
   GtkAdjustment *adj;
-  gint volume;
+  gfloat volume;
   gboolean res = TRUE;
 
   if (!dock->timeout)
